linux/ena: Backport ethtool_puts() conversion
This commit backports [1] which will be part of kernel 6.8.0, a
conversion of few cases of ethtool_sprintf to ethtool_puts.

As explained in [2]:
The first case, with just two arguments:
ethtool_sprintf(&data, buffer[i].name), commonly triggers a
-Wformat-security warning with Clang due to potential problems with
format flags present in the strings [3].

The second, when it's used with format string: "%s"
ethtool_sprintf(&data, "%s", buffer[i].name) is just a bit weird with a
plain-ol' "%s".

This commit also converts the rest of the ethtool_sprintf() calls to
ethtool_puts() for cases when only two arguments are used or when the
call is of the type: ethtool_sprintf(&data, "%s", buffer[i].name);

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git/commit/drivers/net/ethernet/amazon/ena?id=e403cffff1a46ab1a95d3bc72e92634445d68328
[2]: https://lore.kernel.org/lkml/8f0e55ea-1c24-4d6b-9398-0cbc2bb58907@lunn.ch/T/
[3]: https://lore.kernel.org/all/202310101528.9496539BE@keescook/

Signed-off-by: Ofir Tabachnik <ofirt@amazon.com>