linux/ena: Add static to check_cdesc_in_tx_cq
Starting with certain compiler versions, driver
compilation outputs the following warning:

warning: no previous prototype for
‘check_cdesc_in_tx_cq’ [-Wmissing-prototypes]

The reason for this warning is that the function
is used only in that .c file, therefore, it should be
static.

In some configurations, this warning is treated as
an error.

This patch resolves the issue.

Signed-off-by: David Arinzon <darinzon@amazon.com>