linux/ena: Move eth_hw_addr_set to ECC
Due to the increased amount of backports of this
function in different OS distributions, we keep on extending
the kcompat support with more complicated versioning logic,
trying to identify the range of kernel versions of each
distribution where this function is introduced.

Moving this definition to the hands of ECC in order to
avoid the added effort of maintaining the kcompat for
this entry.

Signed-off-by: David Arinzon <darinzon@amazon.com>