linux/ena: Fix redundant device NUMA node override
The driver overrides the NUMA node id of the device regardless of
whether it knows its correct value (often setting it to -1 even though
the node id is advertised in 'struct device'). This can lead to
suboptimal configurations.

This patch fixes this behavior and makes the shared memory allocation
functions use the NUMA node id advertised by the underlying device.

Fixes: 798070ab13c6 ("First commit of Amazon Elastic Network Adapter (ENA) Linux driver")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>