linux/ena: Fix interrupt interval change flag override
Based on feedback in [1], there's a logical mistake in
how interrupt interval change indication flag is being set.
The current logic may cause the following scenario, for
example (Can also happen with DIM being involved):

1. A user runs ethtool, updates RX interval to X.
2. The interrupt interval change indication flag for RX
   is set to true.
3. A user runs ethtool, disables adaptive moderation.
4. No change in interval values has been made, therefore,
   interrupt interval change indication flags for both
   TX and RX are set to false.
5. A user runs ethtool, enables adaptive moderation.
6. No change in interval values has been made, therefore,
   the interrupt interval change indication flags for
   both TX and RX are set to false.
7. NAPI routine executes, the function that updates the
   device with the interrupt moderation information observes
   that both interrupt interval change indication flags are
   set to false, therefore, the device is not updated,
   causing the values to not be propagated to the device and
   the update to be lost.

This patch makes a small change and replaces `=` with `|=`
in order to avoid the override of the current status.
By that, ensuring that there's only one flow that sets
these interrupt interval change indication flags to false
(The logic that checks the values and whether they've been
changed, and update the device).

Fixes: 690b582baae0 ("linux/ena: Add a hint on interrupt moderation update to the device")

[1]: https://lore.kernel.org/netdev/20240506070453.17054-7-darinzon@amazon.com/

Signed-off-by: David Arinzon <darinzon@amazon.com>