fbsd/ena: Rename fields to MBZ in completion descriptors
Several reserved bits in ena_eth_io_tx_cdesc and
ena_eth_io_rx_cdesc_base have been renamed explicitly to
MBZ (Must Be Zero).
These bits are set by the device to zero before being sent
to the driver. The fields are used as an integrity check in
order to ensure that the received descriptor is not corrupted.

in ena_eth_io_tx_cdesc, two reserved bits from the flags
field are marked as MBZ.

in ena_eth_io_rx_cdesc_base, there are 8 reserved bits
(7 and 17-23). Out of the above, bits 7 and 17 are MBZ.

Signed-off-by: David Arinzon <darinzon@amazon.com>