linux/ena: Add a doorbell in XDP error path for ENA_XDP_TX
In case ena_com_rx_pkt() fails while running in the context of
ena_clean_rx_irq(), and at least one packet was queued to TX ring, due
to xdp_verdict which is XDP_TX, we end up having packets in the send
queue without a doorbell being triggered.

This commit fixes this use-case by adding a doorbell if ENA_XDP_TX
verdict is set in the error path.

Fixes: 99770f9af7dc ("linux/ena: Improve XDP_TX throughput by using batching")
Signed-off-by: Evgeny Ostrovsky <evostrov@amazon.com>