linux/ena: Rename variables to more suitable names
In the function ena_xdp_exchange_program_rx_in_range(), the variable
"count" is used to represent the last index of the loop, not the number
of iterations. Therefore, it is more appropriate to name it "last".

Also rename last_monitored_tx_qid. This queue index is also used for
checking rx queue missing interrupts, therefore the current name is
misleading.

Signed-off-by: Evgeny Ostrovsky <evostrov@amazon.com>