linux/ena: Allow LPC for less than 16 channels
LPC is allowed only in case the channel number is no less than 16.
The restriction was added due to LPC's large memory footprint, which may
cause memory issues in smaller instance types.

This commit lifts the limitation and moves the responsibility to the
user. By default, LPC is still disabled for less than 16 channels, but
users will be allowed to enable the feature via a module param or
a private flag setting in ethtool.

LPC can be enabled and tweaked using a module param (a value of 0
disables it). For example:
2. "modprobe ena lpc_size=0".
3. "modprobe ena lpc_size=3".

Signed-off-by: Evgeny Ostrovsky <evostrov@amazon.com>
Signed-off-by: David Arinzon <darinzon@amazon.com>