linux/ena: Use true flexible arrrays in LPC
LWN article [1] advocates that flexible arrays should
be defined with `[]` rather than `[0]` or `[1]` as
the behavior may be unsafe and sometimes undefined
(See [1] for more information).

In addition, a change in GCC 13 (See [2]) changed how
flexible arrays are treated and has increased the
strictness to control which array members are to be
treated as flexible arrays.

When working with Ubuntu 24.04 (GCC 13.2), the
observed issue is that there's an access to an
index out of bounds when trying to access index 1 of
the flexible array used to store the pages in LPC.

The issue doesn't manifest in Ubuntu 22.04 (GCC 11.4).

This patch changes the flexible array from `[0]` to
`[]` as recommended in [1] and is being done in
various locations in the kernel.

[1]: https://lwn.net/Articles/908817/
[2]: https://gcc.gnu.org/gcc-13/changes.html

Fixes: 544fd8759289 ("linux/ena update ENA linux driver to version 2.4.0")
Signed-off-by: David Arinzon <darinzon@amazon.com>