linux/ena: Use tx_info->total_tx_size for validation
When polling for TX completions the driver verifies that the request id
"attached" to this completion represents a valid transmitted packet
(either by checking that it has a pointer to an SKB or an XDP buffer).
The check is done in the function validate_xdp_req_id() or
validate_tx_req_id() (depending on which stack sent the packet).

To allow a unified method for validation for all cases this commit makes
sure 'tx_info->total_tx_size' is non-zero only when we haven't yet
received a completion for it from the device.

This allows to use this field in a future patch for a unified way to
verify valid TX completions.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>