linux/ena: Force ordering on XSK TX queues
AF XDP has no mechanism to report specific TX completions to the
application, but rather only pass the total TX completions that have
been polled. This property requires the driver to ensure that the
transmitted packets are completed in order.

This patch adds logic to the TX completion polling which ensures that
only in-order TX completions are reported.

To help the ordering mechanism, this patch adds an additional flag to
ena_tx_buffer struct. The added field fits inside an existing memory
hole in the struct and doesn't change its overall size or cache
layout.

Fixes: 39c8eee94982 ("linux/ena: Support polling for XSK zero-copy xmit completions")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>