linux/ena: Combine wakeup mechanism for AF XDP TX xmit and poll
This mechanism allows the driver to share its polling state with the AF
XDP program. By clearing the "wakeup needed" flag, the driver tells the
application that it's going to run its napi routine anyway, and doesn't
require a manual trigger.

Since this system is used for both SQ and CQ, this commit combines the
code related to it in both TX transmission and TX completion polling.

This patch also modifies the RX polling function's return value to be
the same as for TX. This is only done for uniformity.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>