linux/ena: Synchronize between AF XDP and regular traffic
The TX traffic sent by the network stack shares the same queue as
packets transmitted by an AF XDP application (XSKs).

This commit adds a lock to synchronize between the two kinds of traffic.

Note that the same variable is also used by XDP_TX/XDP_REDIRECT traffic,
but because different queues are used for this flow, the 'xdp_tx_lock'
is different than the one used by the network-stack/AF XDP.

Fixes: fd80522361cb ("linux/ena: Add AF XDP zero-copy xmit function")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>