linux/ena: Account XDP_PACKET_HEADROOM in UMEM headroom
When asking the driver to assign a UMEM to a queue the size of each
frame's headroom in the UMEM is specified as well.
The requested value is retrieved using the kernel's xsk_pool_get_headroom()
function which returns a *minimum* of XDP_PACKET_HEADROOM [1].

This patch makes the driver to only block headrooms larger than this
minimum (instead of 0).

[1]:
https://github.com/torvalds/linux/blob/e8f897f4afef/include/net/xdp_sock_drv.h#L37

Fixes: 162950f820aa ("linux/ena: Allow setting XSK pools")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>