linux/ena: Make TX req_id be printed by default
For every TX completion the driver verifies whether the request id
number associated to it corresponds to a transmitted packet.
In case of an error the handle_invalid_req_id() function is called.

The error prints in this function use netif_err() with the wrong print
type of 'tx_done' which should be used to give TX completion updates.
Given that the driver doesn't print the 'tx_done' type by default, these
error messages are often silenced.

This commit corrects the type to 'tx_err' which more accurately
describes the prints as errors and ensures that the messages are printed
by default.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>