linux/ena: Always keep ZC TX open
For regular TX traffic the driver can declare a queue as being closed to
signal the network stack that it can process no more TX packets until
some of the previous ones are completed.
This mechanism provides some load balancing on the TX side and improves
performance by allowing traffic to be batched.

The AF XDP (zero-copy) TX path has no such mechanism and the function
polling ZC TX completions doesn't check for the queue state which might
leave a TX queue closed even if it can be re-opened.

This patch prevents the queue from being closed in case the channel is
shared with ZC AF XDP code.

Fixes: fd80522361cb ("linux/ena: Add AF XDP zero-copy xmit function")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>