linux/ena: Exit start_xmit early if queue is full
The non-XDP TX flow (packets arriving from the network stack) runs
ena_start_xmit() with the assumption that there is enough room in the
TX SQ for the packet it wants to transmit.

When AF XDP isn't used this assumption is governed by the
queue-closing mechanism which blocks further transmission after the
driver advertises that it cannot process more packets.

However, when the queue is shared with AF-XDP, the mechanism is
disabled, and the kernel keeps passing packets for transmissions even
when no more entries are available in the SQ.

This commit ensures that TX packets are dropped at the earliest stage
possible if the queue doesn't have enough room. This change ensures
that ena_start_xmit() accesses the TX info data structures only when
they represent an unused packet.

Fixes: fd80522361cb ("linux/ena: Add AF XDP zero-copy xmit function")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>