linux/ena: Restore zero-copy RX buffer offset when recycled
The kernel represents the UMEM buffers for driver's RX side as struct
xdp_buff, with its corresponding attributes:
- xdp_buff->data_hard_start: the beginning of the RX buffer
- xdp_buff->data: the beginning of the packet inside the buffer
- xdp_buff->data_end: the end of the packet inside the buffer

When the driver allocates an RX buffer for its zero-copy queue (using
xsk_buff_alloc()) the kernel fetches a buffer from the UMEM and
initializes the fields mentioned above.

In case the received packet contains an offset added to it by
the device, xdp_buff->data is updated with this offset.

This patch ensures that in case the RX buffer is recycled (e.g. if
the XDP program decides to drop the packet) the value of xdp_buff->data
is still correct even when there's an offset from the device.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>