linux/ena: Introduce support for flow steering
Flow steering is the ability to give the user the option to steer
RX traffic to specific queues, by noting specific fields and values
in the packets headers.

This commit contains the initial support for this feature.
This includes the resource allocation and communication with the device
to retrieve its support capabilities, and subsequently conveying this
information to the OS.

Rules are re-configured from driver memory back to the device in
case of an internal error that caused a reset.

The patch adds support to ethtool interface to add, remove, and fetch
the rules configured.

Signed-off-by: Sun Amar <sunamar@amazon.com>