linux/ena: Fix function hiding that causes XDP not to work properly
Commit [1] defines the functions [2] in kernels < 6.3 as empty macros.
Some distributions (i.e. RHEL 9) backported these functions to
kernels < 6.3.
In such kernels the functions are hidden by the empty macros, disabling
their implementation.

This can cause XDP not to work properly, since the xdp_features are no
longer set/cleared as intended by the code.

For example in RHEL 9 this causes XDP_REDIRECT not to work.
It happens because:
1. When ena_xdp_set() runs it calls xdp_features_set_redirect_target()
2. xdp_features_set_redirect_target() sets the NETDEV_XDP_ACT_NDO_XMIT
   flag in its implementation, but because our code hides the
   implementation, the flag is not set.
3. xdp_do_redirect() calls sub-functions until it gets to
   __xdp_enqueue(), which returns -EOPNOTSUPP if NETDEV_XDP_ACT_NDO_XMIT
   is not set.

This commit:
1. Uses ECC to figure out if functions [2] are defined or not.
2. Uses ECC to make sure that xdp_features is defined in struct
   net_device before setting them.

[1]: See Fixes
[2]: xdp_features_set_redirect_target()
     xdp_features_clear_redirect_target()
     xdp_set_features_flag()
     xdp_clear_features_flag()

Fixes: 58ba634a1aa8 ("linux/ena: Backport XDP features netlink advertisement")
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>