linux/ena: Support PHC when adjfine is not implemented
Commit [1] introduced the adjfine() hook for PTP clocks.
Patchset [2] represents a transition from the adjfreq()
hook to the adjfine() hook, which included a conversion
of all the drivers adjfreq() to adjfine().
As a result, in pathset [2], there's an implicit assumption
that all drivers implement adjfine(), without a check
of whether the hook has been added.
However, our driver doesn't implement adjfine(), and PHC wasn't
part of the kernel during the above mentioned effort, leaving
this support missing.
This results in an assertion error, which was reported in [3].

This commit adds the adjfine() hook.

[1]: https://github.com/torvalds/linux/commit/d8d26354191399627bac9cf0da0667b0f5178686
[2]: https://lore.kernel.org/netdev/20221109230945.545440-1-jacob.e.keller@intel.com/
[3]: https://github.com/amzn/amzn-drivers/issues/322

Signed-off-by: David Arinzon <darinzon@amazon.com>