linux/ena: Update non-adaptive interrupt interval for XDP queues
non-adaptive interrupt interval can be updated via ethtool
for all the driver queues.
When an XDP program is loaded, the driver queues are split
in the following way:
1. [0, num_io_queues) - I/O queues
2. [num_io_queues, num_io_queues + num_xdp_queues) - XDP queues

Currently, the logic that updates the non-adaptive value
only updates it for the I/O queues, leaving the XDP queues
with the old value.

This patch takes the XDP queues into account.

Fixes: 690b582baae0 ("linux/ena: Add a hint on interrupt moderation update to the device")

Signed-off-by: David Arinzon <darinzon@amazon.com>