linux/ena: Don't complete napi in XDP if budget is consumed
In case the budget is consumed, it means that there are
probably more packets to be processed, therefore, the
budget value is sent back to the napi poll caller,
indicating that another scheduling is required.

In the current code, if `ENA_AF_XDP_SUPPORT` is enabled,
in `ena_xdp_io_poll()`, the work done by the XDP
routine `ena_clean_xdp_irq()` is not taken into account
and `needs_wakeup` is set to a default true, causing
napi to be marked as complete and interrupts unmasked.
After all of this, the napi caller function may
receive the `budget` value as a return, indicating the
mentioned above (re-scheduling).

This scenario creates a case where napi is marked as
completed, while the `budget` return indicates that
napi schedule is required. In parallel to this,
an interrupt is triggered and schedules napi as well
with `napi_schedule_irqoff()`, causing a double add
of the same napi to the napi list maintained by the
kernel core code.

The above scenario leads to a kernel panic like the below

[ 3716.003483]  ? show_trace_log_lvl+0x1c4/0x2d2
[ 3716.003854]  ? show_trace_log_lvl+0x1c4/0x2d2
[ 3716.004219]  ? __napi_schedule_irqoff+0x58/0x80
[ 3716.004598]  ? __die_body.cold+0x8/0xd
[ 3716.004907]  ? die+0x2b/0x50
[ 3716.005160]  ? do_trap+0xce/0x120
[ 3716.005447]  ? __list_add_valid.cold+0x23/0x5b
[ 3716.005822]  ? do_error_trap+0x65/0x80
[ 3716.006137]  ? __list_add_valid.cold+0x23/0x5b
[ 3716.006507]  ? exc_invalid_op+0x4e/0x70
[ 3716.006826]  ? __list_add_valid.cold+0x23/0x5b
[ 3716.007202]  ? asm_exc_invalid_op+0x16/0x20
[ 3716.007540]  ? __list_add_valid.cold+0x23/0x5b
[ 3716.007913]  ? __list_add_valid.cold+0x23/0x5b
[ 3716.008285]  __napi_schedule_irqoff+0x58/0x80
[ 3716.008650]  ena_intr_msix_io+0x37/0x50 [ena]
[ 3716.009021]  __handle_irq_event_percpu+0x43/0x180
[ 3716.009416]  handle_irq_event+0x34/0x70
[ 3716.009740]  handle_edge_irq+0x87/0x220
[ 3716.010067]  __common_interrupt+0x3b/0xa0
[ 3716.010408]  common_interrupt+0x7b/0xa0

This patch takes the work done by the XDP function
`ena_clean_xdp_irq()` into account when checking whether
wakeup is required (which means that napi is truly done, and
the budget is not fully consumed).

Fixes: 3575da07847f ("linux/ena: Introduce ZC queue TX completion polling")
Signed-off-by: David Arinzon <darinzon@amazon.com>