linux/ena: Backport RX software timestamp set for all
In patch [1], `SOF_TIMESTAMPING_RX_SOFTWARE` and
`SOF_TIMESTAMPING_SOFTWARE` were moved to `__ethtool_get_ts_info`,
therefore, drivers no longer need to report these when the
ethtool `get_ts_info` hook is called.

The mentioned patch has been introduced to net-next and, at
the time of this commit, intended for upstream kernel 6.12.
Even though is not a function addition/removal/change,
and it is difficult to identify whether the change is in or not
during runtime for a given kernel, added an `#if` check for
kernel 6.12 to allow better clarity.
Having the addition of `SOF_TIMESTAMPING_RX_SOFTWARE` and
`SOF_TIMESTAMPING_SOFTWARE` even though change [1] is in
will not cause a negative effect.

[1]: https://lore.kernel.org/netdev/IA1PR12MB6113E1D2165FB9F87F8DA06DCE902@IA1PR12MB6113.namprd12.prod.outlook.com/T/

Signed-off-by: David Arinzon <darinzon@amazon.com>