linux/ena: Don't allow ethtool to choose flow steering rule location
ethtool allows the option to insert a flow steering rule without
a specific location, by not specifying `loc` in `ethtool -N/U`
command.
In such a case, ethtool queries the number of current and total
rules using `ETHTOOL_GRXCLSRLCNT` and chooses the highest (lowest
priority) available index.
As the flow steering table is shared between all interfaces of
an instance, there can be a collision and one of the interfaces
will get a failure to add.
Adding `RX_CLS_LOC_SPECIAL` in `ETHTOOL_GRXCLSRLCNT` command
`data` field indicates to ethtool to not choose the location
by itself.

ENA recommends setting rules with `loc`, the locations need
to be managed by the user.

Signed-off-by: David Arinzon <darinzon@amazon.com>