linux/ena: Add ndo_rx_flow_steer hook
With the introduction of n-tuple support in [1], `NETIF_F_NTUPLE`
is advertised on supported devices.
The kernel assumes that if `NETIF_F_NTUPLE` is set, then
`ndo_rx_flow_steer` is implemented (See the code in [2]).
As ENA doesn't support `ndo_rx_flow_steer`, add the hook
with an unsupported return value in order to avoid a kernel panic
with NULL pointer dereference. This may happen when
the user configures RPS and tries to run network traffic.

[1]: See Fixes tag
[2]: https://github.com/torvalds/linux/blob/master/net/core/dev.c#L4556

Fixes: 0d5aca93bcfa ("linux/ena: Introduce support for flow steering")
Signed-off-by: David Arinzon <darinzon@amazon.com>