linux/ena: Don't recycle redirected buffers on failure
When using AF XDP, the eBPF program passes packets to the user-space
application through the XDP_REDIRECT return value.
The driver then calls `xdp_do_redirect` to perform the redirection. In
case the userspace application doesn't have enough room in its RX CQ,
the redirection would fail.

Starting version 6.8 (commit 269009893146c [1]), when the redirected
buffer belongs to a UMEM buffer (zero-copy mode) the kernel takes care
to "free" the buffer and re-enqueues it into the FILL ring. This
essentially breaks the ownership context since it's the driver that
allocated the memory.

As noticed by a Github customer [2], when xdp_do_redirect() fails, the
driver, that is convinced that the buffer is in its ownership, re-uses
the same buffer when submitting an RX descriptor to the device.
Since the buffer was *also* recycled by the kernel, once the driver
allocates a UMEM buffer, the kernel returns the same address that was
already recycled.
This creates a memory corruption due to simultaneous writes to the same
buffer.

To avoid guessing whether the kernel recycled the buffer on failure and
whether zero-copy is in use, this patch makes all redirection errors
remove driver ownership of the buffer and not recycle it.

*Note* that the driver calls xdp_return_buff() unconditionally:
    - For non AF XDP traffic, the kernel doesn't try to free the buffer
      as well, and so it's safe to free it in the driver
    - For AF XDP copy mode, the buffer is also not freed by the kernel
      and it's safe to free it in the driver
    - For AF XDP zero-copy mode, even if the kernel freed the buffer already,
      starting version 5.16 (commit [3]) the kernel has a protection
      against double free, and so it should be safe.
      There is a theoretical risk that commit [1] would be backported
      but [3] wouldn't but this seems unlikely.

kcompat entry:
Unfortunately the generic xdp_return_buff() which can handle the free
operations for both UMEM memory and page-pool pages has been exported
only in kernel 5.18 [4].
Since the function existed before that, but only changed its linking
status, ECC cannot be used to detect it since it doesn't do linking
phase.

There is an edge case in which the kcompat implementation doesn't
succeed: If the conversion of an xdp_buff to xdp_frame fails, the driver
has no other option but to give up and not free the buffer and recycle
it.

[1]: 269009893146 ("xsk: recycle buffer in case Rx queue was full")
https://github.com/torvalds/linux/commit/269009893146c
[2]: https://github.com/amzn/amzn-drivers/issues/330
[3]: 199d983bc015 ("xsk: Fix crash on double free in buffer pool")
https://github.com/torvalds/linux/commit/199d983bc0151
[4]: 718a18a0c8a6 ("veth: Rework veth_xdp_rcv_skb in order to accept non-linear skb")
https://github.com/torvalds/linux/commit/718a18a0c8a67

Fixes: 0f361baa6286 ("linux/ena: Add AF XDP zero-copy RX polling function")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>