linux/ena: Backport pass dim_sample to net_dim() by reference
This is a backport of commit [1] from the community.
The change is part of kernel 6.13.

The change was also done to our dim files because there
might be a case where `ENA_NET_DIM_SAMPLE_PARAM_BY_REF`
is defined as the change is backported, but the kernel
is lower than version 5.3.0 and our dim files will
be compiled, but with the previous signature (call
by value), causing a compilation error.

[1]: https://lore.kernel.org/netdev/20241030164805.40408945@kernel.org/T/

Signed-off-by: David Arinzon <darinzon@amazon.com>
Signed-off-by: Caleb Sander Mateos <csander@purestorage.com>