linux/ena: Prevent crash from uninitialized ena_dev in admin queue
The ena_dev field inside the admin queue structure is initialized at
the end of the admin queue initialization process.
This late initialization may cause a system crash if ena_trc_err
is called during the initialization of different admin queue components.
In such a case, the uninitialized ena_dev field will be used,
leading to the crash.

Fixes: a803fb53d633 ("linux/ena update ENA linux driver to version 2.2.11")
Signed-off-by: Wajeeh Atrash <atrwajee@amazon.com>