linux/ena: Account for all bytes of a recycled XDP packet
When an XDP program is loaded and multi-buffer packets arrive
ENA_XDP_RECYCLE is returned.

In such a case the packet is counted in the packets stat
but only the first buffer of the packet is accounted for
in the bytes stat.

This commit fixes this by accounting for all buffers
of the packet in the bytes stat.

Fixes: 51310a616aa6 ("linux/ena: Fix several bugs in XDP infrastructure")
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>