linux/ena: Account for all bytes of a recycled AF_XDP packet
When a multi-buffer packets arrive during an AF_XDP ZC scenario
ENA_XDP_RECYCLE is returned.

In such a case the packet is counted in the packets stat
but only the first buffer of the packet is accounted for
in the bytes stat.

This commit fixes this by accounting for all buffers
of the packet in the bytes stat.

Fixes: 0f361baa6286 ("linux/ena: Add AF XDP zero-copy RX polling function")
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>