linux/ena: Remove unnecessary split of ENA_XDP_MAX_MTU to cases
Commit [1] unnecessarily split ENA_XDP_MAX_MTU to 2 cases according to
XDP_HAS_FRAME_SZ definition, without explaning why it was done.

When an XDP program returns XDP_PASS on a received packet, build_skb()
is called.
build_skb() initializes the shared_info struct at the tail of the
first buffer. Therefore it is necessary to reserve
SKB_DATA_ALIGN(sizeof(struct skb_shared_info)) in the rx buffer.
This is true whether XDP_HAS_FRAME_SZ is defined or not.

Currently the code subtracts the size of the Ethernet FCS field,
(a checksum field used to detect corruption), however this field
is not written by the ENA device in the RX buffer, therefore there
is no need to take it into account in the MTU calculation.

[1]: See 'Fixes' tag

Fixes: a803fb53d633 ("linux/ena update ENA linux driver to version 2.2.11")
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>