linux/ena: Add check if XDP multi-buffer is supported using ECC
Check with ECC if all members and functions necessary for XDP
multi-buffer support are present, and define ENA_XDP_MB_SUPPORTED
accordingly.

All the different members needed for XDP multi-buffer support are
grouped into one ECC entry to reduce the increase of compilation
time as much as possible.

The only outlier is skb_frag_fill_page_desc() which is not strictly
necessary for XDP multi-buffer support, however it was decided
to use it in the xdp multi-buffer path, and therefore it is added
where missing.

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>