linux/ena: Backport utilization of netif_enable_cpu_rmap
This is a backport of [1], which was added to net-next as
a candidate for Upstream Kernel 6.15.

A partial quote from the patch:

Move the IRQ affinity management to the napi struct. This way we can
have a unified IRQ notifier to re-apply the user-set affinity and
also manage the ARFS rmaps.

The first patch moves the aRFS rmap management to CORE. It also adds
the IRQ affinity mask to napi_config and re-applies the mask after
reset.

[1]: https://lore.kernel.org/netdev/20250224232228.990783-4-ahmed.zaki@intel.com/T/

Signed-off-by: David Arinzon <darinzon@amazon.com>