linux/ena: resolve WARN_ON when freeing IRQs
Change [1] changed the method of how NAPIs and IRQs are
associated and managed.
The change introduced an issue where a WARN_ON in [2] is
triggered as the affinity notifier is not released.
The above results in the below stack trace when freeing
the IRQs:

[  484.544586]  ? __warn+0x84/0x130
[  484.544843]  ? free_irq+0x5c/0x70
[  484.545105]  ? report_bug+0x18a/0x1a0
[  484.545390]  ? handle_bug+0x53/0x90
[  484.545664]  ? exc_invalid_op+0x14/0x70
[  484.545959]  ? asm_exc_invalid_op+0x16/0x20
[  484.546279]  ? free_irq+0x5c/0x70
[  484.546545]  ? free_irq+0x10/0x70
[  484.546807]  ena_free_io_irq+0x5f/0x70 [ena]
[  484.547138]  ena_down+0x250/0x3e0 [ena]
[  484.547435]  ena_destroy_device+0x118/0x150 [ena]
[  484.547796]  __ena_shutoff+0x5a/0xe0 [ena]
[  484.548110]  pci_device_remove+0x3b/0xb0
[  484.548412]  device_release_driver_internal+0x193/0x200
[  484.548804]  driver_detach+0x44/0x90
[  484.549084]  bus_remove_driver+0x69/0xf0
[  484.549386]  pci_unregister_driver+0x2a/0xb0
[  484.549717]  ena_cleanup+0xc/0x130 [ena]
[  484.550021]  __do_sys_delete_module.constprop.0+0x176/0x310
[  484.550438]  ? syscall_trace_enter+0xfb/0x1c0
[  484.550782]  do_syscall_64+0x5b/0x170
[  484.551067]  entry_SYSCALL_64_after_hwframe+0x76/0x7e

Based on patchset [1], adding a call to `netif_napi_set_irq`
with -1 as the IRQ index, which frees the notifier.

The issue is re-created when doing `modprobe -r ena; modprobe ena`,
but can be also re-created in any reconfiguration where I/O queues
are freed and re-created (as well as their IRQ associations).

The issue was observed in net-next (Upstream Kernel 6.15 candidate)

[1]: See Fixes tag
[2]: https://web.git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git/tree/kernel/irq/manage.c#n2031

Fixes: https://lore.kernel.org/netdev/20250224232228.990783-4-ahmed.zaki@intel.com/T/

Signed-off-by: David Arinzon <darinzon@amazon.com>