fbsd/ena: Fix misconfiguration when requesting regular LLQ
Patch [1] introduced new values to hw.ena.force_large_llq_header.
The default value of 2 means no preference, while 0 and 1 act as the
previous false and true respectively, which allowed forcefully setting
regular or large LLQ.

There are 2 ways to force the driver to select regular LLQ:

1. Setting hw.ena.force_large_llq_header = 0 via sysctl.
2. Turning on ena express.

When the device supports large LLQ but the driver is forced to
regular LLQ, llq_config->llq_ring_entry_size_value is never initialized
and since it is a variable allocated on the stack, it stays garbage.

Since this variable is involved in calculating max_entries_in_tx_burst,
it could cause the maximum burst size to be zero. This causes the driver
to ignore the real maximum burst size of the device, leading to driver
resets in devices that have a maximum burst size (Nitro v4 and on. see
[2] for more information).

In case the garbage value is 0, the calculation of
max_entries_in_tx_burst divides by 0 and causes kernel panic.

The patch modifies the logic to take into account all use-cases and
ensure that the relevant fields are properly initialized.

[1]: See Fixes
[2]: https://docs.aws.amazon.com/ec2/latest/instancetypes/ec2-nitro-instances.html

Fixes: d870eb0e6761 ([New Feature] Support LLQ entry size recommendation from device)
Signed-off-by: David Arinzon <darinzon@amazon.com>
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>