linux/ena: Avoid setting xdp_ring pointer with uninitialized rings
Background and issue description:
---------------------------------
The ena driver allows loading an XDP program only if the number of
queues is set to be (at most) half of the maximum allowed. This is
because it adds a second half of tx queues to be used for XDP tx
traffic.

The xdp_ring field in ena_ring was added to allow quicker access to
the relevant xdp_ring when needed.

The xdp_ring for rx_ring `i` is initialized to be tx_ring
`i + num_io_queues`.

This works well when the number of queues is at most half of the
maximum, because each rx ring can point to the relevant tx ring in the
second half of the rings. However, when the number of rings is larger,
xdp_ring points to memory outside of the tx rings.

This becomes an issue if the device supports more than 64 queues
(and up to 128), since xdp_ring will point to queue indexes between
128-255, which are not inside the memory initialized for io queues.
In this case if UBSAN (see [1]) is turned on, it will print a warning
and call stack in dmesg during queue initialization.

This commit:
------------
Doesn't initialize the xdp_ring field for queues, whose xdp_ring
will overflow the max_num_io_queues.

[1]: https://docs.kernel.org/dev-tools/ubsan.html

Fixes: 4e0c78897589 ("linux/ena: a few minor performance improvements")
Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>