linux/ena: Synchronize AQ with the interrupt on timeout
The admin queues consist of an SQ to submit a command to the device and
a CQ to read its response.
Both queues share an object called 'command_ctx' which is a struct
containing the details of a submitted command and (after completing the
request) its completion status.

The flow works as follows:
- Driver submits an entry to admin SQ with a cmd id which identifies a
  specific command context
- Device submits an entry to the admin CQ with the same cmd id and command
  status (success/failure).

When the driver polls the CQ for completions, the command context is
fetched by the attached cmd id and it is updated with the status of the
command.

The driver can update the command context in two ways:
- polling mode: after submitting the request the same thread polls the
  CQ for the device response until the command succeeds or a timeout is
  reached. The thread sleeps in increasing intervals between each poll.

- interrupt mode: after submitting the request the process moves to
  sleep, until an interrupt (admin IRQ) is received and signals the
  process to wake, or until a timeout is reached.

In interrupt mode, if the process submitting the request times out
before receiving a signal, it starts polling the admin queue by itself
to analyze the reason for the failure.

This creates a race in which two threads (the submitter thread and the
interrupt) access shared resources. To name a few bugs arising from this
shared access:
1. If the submitter thread polls the admin queue and decides to free the
   resources attached to an entry (e.g. user_cqe field), the interrupt
   thread might end up accessing stale or NULL data.
2. Both threads might modify the queue head counter making its value
   undefined

This patch also ensures that the final status of the command (held in
the 'status' member) is only updated once the processing of the command
completion is over.

Fixes: 798070ab13c6 ("First commit of Amazon Elastic Network Adapter (ENA) Linux driver")
Signed-off-by: Shay Agroskin <shayagr@amazon.com>