linux/ena: Move last_napi_jiffies to the beginning of NAPI
`last_napi_jiffies` is used to indicate when was the last
time that the NAPI routine has been executed.
A high value may indicate that NAPI has been starved or
not scheduled.
There can be use-cases where NAPI has been scheduled but
the NAPI routine itself may take time and is a subject
to preemption.

This patch moves `last_napi_jiffies` to the beginning of
the NAPI routine to indicate more clearly when NAPI
scheduling actually happened without taking into account
the amount of time the NAPI routine executed.

Signed-off-by: David Arinzon <darinzon@amazon.com>