linux/ena: HW packet timestamping support
Add support for retrieving HW timestamps for TX and RX packets.
The feature is configurable through ioctls and the timestamps
can be retrieved by using the relevant socket operations.

A documentation entry has been added to provide additional
information (and a reference to [1]) as well as an
alternative method of retrieving the configuration through
sysfs.

[1]: https://www.kernel.org/doc/html/latest/networking/timestamping.html

Signed-off-by: David Arinzon <darinzon@amazon.com>