linux/ena: Correct Identification of Missing Completions in Device
Before reviewing this patch, we need to define two key concepts:

1. What constitutes a missing TX completion reset?
2. What constitutes a missing interrupt reset?

The challenge is that the driver will experience similar symptoms
in both cases, specifically pending TX descriptors in the Submission
Queue (SQ). Our goal is to differentiate between the two.

We define the concepts as follows:

1. Missing TX completion: This refers to the case where the device
   fails to send completions to the driver, meaning there are missing
   completion descriptors in the driver's Completion Queue (CQ).
2. Missing interrupts: This refers to the case where the device
   sends the completions, but the interrupt is lost on the way to the
   driver.

Note that from the driver's perspective, both scenarios exhibit similar
symptoms — completions pending in the TX SQ. One way to identify the
root cause of the issue is to inspect the CQ and check whether there
are completions corresponding to the timed-out descriptors.

Issue Description:
------------------
When completions are lost in the device and the driver does not receive
completions for the sent descriptors, the current code checks the CQ for
any descriptors. If it finds a descriptor, it assumes that the
descriptor belongs to an expired sent descriptor from the SQ and
issues a reset reason indicating a missing interrupt. This assumption
may be incorrect.
Furthermore, we trigger reset when there are more than missed_tx_thresh
timed out pending tx submissions but it would be more accurate to check
against the value of currently pending tx pkts that didn't get their
completion processed yet, this value is represented by
pending_timedout_pkts statistic.
Also, not all cases of missing first interrupt are correctly
classified, when we hit missed_tx_thresh we don't check if we received
first interrupt.

Fix:
----
To resolve this issue, we need to iterate over all entries in the CQ
to verify whether the completions correspond to expired descriptors.
Also, we need to verify if we have received first interrupt when
missed_tx_thresh is reached.
For comparing against missed_tx_thresh we will use
pending_timedout_pkts which represents a more accurate number of timed
out tx pkts that didn't get their completion processed yet.
This change improves the debuggability and correctness of the system and
provides more insight into the state of the CQ when issues arise.

Signed-off-by: Osama Abboud <osamaabb@amazon.com>
Signed-off-by: Ofir Tabachnik <ofirt@amazon.com>